HumanTimeBasedUpdate
-TimeBased Update rework

+++ Assets/Script/Humans/HumansStats/HumansStats.cs
@@ -9,6 +9,6 @@ public class HumansStats : ScriptableObject
     public int currentAge = 18;
     public int lifeEspectancy = 110;
     public int strenght = 60;
-    public int speed = 20;
+    public float speed = 20;
 
 }
